fix: type field filter employee

+++ src/main/java/com/rgv04/hr/domain/employee/EmployeeFilter.java
@@ -10,9 +10,9 @@ import lombok.Setter;
 public class EmployeeFilter {
 
     @NotNull
-    private int startYear;
+    private Long startYear;
 
     @NotNull
-    private int endYear;
+    private Long endYear;
 
 }
